pyproject.toml
[project]
name = "cua_template"
version = "0.1.0"
description = "A browser automation agent using Tzafon's computer use API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
    "tzafon>=2.8.1",
]

[tool.setuptools.packages.find]
where = ["src"]
